KETTLE-46:  Move `kettle.dataSource.URL` to its own file to make it easier to use it in-browser.
